Fix implicit casts

+++ src/bson_array.c
@@ -176,7 +176,7 @@ BsonArray bson_array_from_bytes(uint8_t *data) {
         //String length is read first
         int32_t stringLength = read_int32_le(&current) - 1;
 
-        char *stringVal = byte_array_to_bson_string(current, stringLength);
+        char *stringVal = byte_array_to_bson_string(current, (size_t)stringLength);
         bson_array_add_string(&array, stringVal);
         free(stringVal);
         current += stringLength + 1;
